working funtionalities for add and remove
