+++ src/style.css
@@ -78,7 +78,7 @@ input[type=text] {
 button {
   border: 0;
   width: 100%;
-  height: 8vw;
+  height: 4vw;
   font-weight: bold;
 }
 
@@ -117,6 +117,5 @@ label:hover {
 }
 
 .hide {
-display: none;
+  display: none;
 }
-
